restructure to share common resources
