+++ Cargo.toml
@@ -31,11 +31,14 @@ base64ct.workspace = true
 bitvec = "1.0.1"
 bytes = "1.10.1"
 chrono = { version = "0.4.41", features = ["serde"] }
-ciborium.workspace = true
+ciborium = "0.2.2"
+cid = { version = "0.11.1", features = ["alloc"] }
 coset = "0.3.8"
 credibil-identity.workspace = true
 credibil-jose.workspace = true
 derive_builder = "0.20.2"
+ecdsa = "0.16.9"
+ed25519-dalek = { version = "2.1.1", features = ["rand_core"] }
 fastrand = "2.3.0"
 flate2 = "1.1.1"
 form_urlencoded = "1.2.1"
@@ -44,6 +47,8 @@ http = "1.3.1"
 http-body = "1.0.1"
 http-body-util = "0.1"
 image = { version = "0.25.6", default-features = false, features = ["png"] }
+k256 = "0.13.4"
+multihash-codetable = { version = "0.1.4", features = ["sha2"] }
 percent-encoding = "2.3.1"
 qrcode = { version = "0.14.1", default-features = false, features = ["image"] }
 rand = "0.9.1"
@@ -55,15 +60,10 @@ thiserror = "2.0.12"
 tracing = "0.1.41"
 uuid = { version = "1.16.0", features = ["fast-rng", "v4"] }
 
-ecdsa = "0.16.9"
-ed25519-dalek = { version = "2.1.1", features = ["rand_core"] }
-k256 = "0.13.4"
-
 [dev-dependencies]
 axum = { version = "0.8.4", features = ["macros"] }
 axum-extra = { version = "0.10.1", features = ["typed-header"] }
 ed25519-dalek.workspace = true
-
 hex = "0.4.3"
 insta = { version = "1.43.1", features = ["redactions", "yaml"] }
 oauth2 = { version = "5.0.0" }
@@ -116,12 +116,12 @@ nursery = "warn"
 [workspace.dependencies]
 anyhow = "1.0.98"
 base64ct = { version = "1.7.3", features = ["std"] }
-ciborium = "0.2.2"
 credibil-identity = { git = "https://github.com/credibil/identity.git" }
 credibil-jose = { git = "https://github.com/credibil/jose.git" }
 credibil-vc = { path = "." }
 ed25519-dalek = "2.1.1"
 futures = "0.3.31"
+
 rand_core = "0.6.4"
 serde = { version = "1.0.219", features = ["derive"] }
 serde_json = { version = "1.0.140", features = ["alloc"] }
