+++ crates/provider/Cargo.toml
@@ -11,15 +11,10 @@ crate-type = ["cdylib", "rlib"]
 anyhow = "1.0.98"
 base64ct.workspace = true
 blockstore = "0.7.1"
-ciborium.workspace = true
-cid = { version = "0.11.1", features = ["alloc"] }
 credibil-identity.workspace = true
 credibil-jose.workspace = true
 credibil-vc.workspace = true
 ed25519-dalek.workspace = true
-# futures.workspace = true
-multihash-codetable = { version = "0.1.4", features = ["sha2"] }
 rand_core.workspace = true
-serde = { version = "1.0.219", features = ["derive"] }
 serde_json.workspace = true
 x25519-dalek = { version = "2.0.1", features = ["static_secrets"] }
