+++ crates/provider/src/identity.rs
@@ -11,7 +11,7 @@ use credibil_identity::did::{
 };
 use credibil_identity::{Identity, IdentityResolver, Key, SignerExt};
 use credibil_jose::{Algorithm, Curve, KeyType, PublicKeyJwk};
-use credibil_vc::core::generate;
+use credibil_vc::generate;
 
 use crate::blockstore::Mockstore;
 use crate::keystore::{self, KeyUse, Keyring, SigningKey};
