+++ crates/provider/src/issuer.rs
@@ -4,7 +4,7 @@ use anyhow::Result;
 use credibil_identity::did::Document;
 use credibil_identity::{Identity, IdentityResolver, Key, SignerExt};
 use credibil_jose::{Algorithm, Signer};
-use credibil_vc::BlockStore;
+use credibil_vc::blockstore::BlockStore;
 
 use crate::blockstore::Mockstore;
 use crate::identity::DidIdentity;
