+++ crates/provider/src/verifier.rs
@@ -4,7 +4,7 @@ use anyhow::Result;
 use credibil_identity::{Identity, IdentityResolver, Key, SignerExt};
 use credibil_jose::jwe::{PublicKey, SharedSecret};
 use credibil_jose::{Algorithm, Receiver, Signer};
-use credibil_vc::BlockStore;
+use credibil_vc::blockstore::BlockStore;
 use credibil_vc::token_status::StatusToken;
 
 use crate::blockstore::Mockstore;
