+++ crates/provider/src/wallet.rs
@@ -6,12 +6,12 @@ use anyhow::Result;
 use base64ct::{Base64UrlUnpadded, Encoding};
 use credibil_identity::{Identity, IdentityResolver, Key, SignerExt};
 use credibil_jose::{Algorithm, Jws, Signer};
-use credibil_vc::core::serde_cbor;
 use credibil_vc::format::FormatProfile;
 use credibil_vc::format::mso_mdoc::{IssuerSigned, MobileSecurityObject};
 use credibil_vc::format::sd_jwt::SdJwtClaims;
 use credibil_vc::oid4vci::types::Credential;
 use credibil_vc::oid4vp::types::{Claim, Queryable};
+use credibil_vc::serde_cbor;
 use serde_json::Value;
 
 use crate::blockstore::Mockstore;
