+++ examples/issuer.rs
@@ -14,14 +14,15 @@ use axum::{Form, Json, Router};
 use axum_extra::TypedHeader;
 use axum_extra::headers::authorization::Bearer;
 use axum_extra::headers::{Authorization, Host};
-use credibil_vc::core::http::IntoHttp;
+use credibil_vc::blockstore::BlockStore;
+use credibil_vc::http::IntoHttp;
 use credibil_vc::oid4vci::{
     self, AuthorizationRequest, CreateOfferRequest, CredentialHeaders, CredentialOfferRequest,
     CredentialRequest, DeferredCredentialRequest, IssuerRequest, NotificationHeaders,
     NotificationRequest, PushedAuthorizationRequest, ServerRequest, TokenRequest,
 };
 use credibil_vc::token_status::StatusListRequest;
-use credibil_vc::{BlockStore, urlencode};
+use credibil_vc::urlencode;
 use oauth2::CsrfToken;
 use provider::issuer::data::{CLIENT, ISSUER, NORMAL_USER as USER, SERVER};
 use provider::issuer::{BOB_ID, ISSUER_ID, Issuer};
