+++ examples/verifier.rs
@@ -11,8 +11,8 @@ use axum::routing::{get, post};
 use axum::{Form, Json, Router};
 use axum_extra::TypedHeader;
 use axum_extra::headers::Host;
-use credibil_vc::BlockStore;
-use credibil_vc::core::http::IntoHttp;
+use credibil_vc::blockstore::BlockStore;
+use credibil_vc::http::IntoHttp;
 use credibil_vc::oid4vp::{self, AuthorzationResponse, GenerateRequest, RequestUriRequest};
 use provider::verifier::data::VERIFIER;
 use provider::verifier::{VERIFIER_ID, Verifier};
