+++ src/core.rs
@@ -4,10 +4,11 @@
 // // https://github.com/rust-lang/rust/issues/99301
 // #![feature(error_generic_member_access)]
 
+pub mod blockstore;
 pub mod generate;
 pub mod http;
-pub mod pkce;
 pub mod serde_cbor;
+pub mod state;
 pub mod strings;
 pub mod urlencode;
