+++ src/lib.rs
@@ -16,11 +16,12 @@ pub mod oid4vci;
 #[cfg(feature = "verifier")]
 pub mod oid4vp;
 
-pub mod core;
-mod endpoint;
 pub mod format;
 pub mod oauth;
 pub mod token_status;
+
+mod core;
+mod endpoint;
 
 /// Re-export DID resolution
 pub mod identity {
@@ -33,29 +34,4 @@ pub mod jose {
 }
 
 /// Re-export basic types
-pub use crate::core::{Kind, OneMany, urlencode};
-
-/// `BlockStore` is used by implementers to provide data storage
-/// capability.
-pub trait BlockStore: Sized + Send + Sync {
-    /// Store a data block in the underlying block store.
-    fn put(
-        &self, owner: &str, partition: &str, cid: &str, data: &[u8],
-    ) -> impl Future<Output = anyhow::Result<()>> + Send;
-
-    /// Fetches a single block by CID from the underlying store, returning
-    /// `None` if no match was found.
-    fn get(
-        &self, owner: &str, partition: &str, cid: &str,
-    ) -> impl Future<Output = anyhow::Result<Option<Vec<u8>>>> + Send;
-
-    /// Delete the data block associated with the specified CID.
-    fn delete(
-        &self, owner: &str, partition: &str, cid: &str,
-    ) -> impl Future<Output = anyhow::Result<()>> + Send;
-
-    /// Purge all blocks from the store.
-    fn purge(
-        &self, owner: &str, partition: &str,
-    ) -> impl Future<Output = anyhow::Result<()>> + Send;
-}
+pub use crate::core::*;
