+++ src/oid4vci/server/token.rs
@@ -17,9 +17,10 @@ use std::fmt::Debug;
 
 use anyhow::Context as _;
 
-use crate::core::{generate, pkce};
+use crate::core::generate;
 use crate::oauth::GrantType;
 use crate::oid4vci::endpoint::{Body, Error, Handler, Request, Response, Result};
+use crate::oid4vci::pkce;
 use crate::oid4vci::provider::{Metadata, Provider, StateStore};
 use crate::oid4vci::state::{Expire, Stage, State, Token};
 use crate::oid4vci::types::{
