+++ src/oid4vp.rs
@@ -161,15 +161,14 @@
 //! [OpenID4VP]: (https://openid.net/specs/openid-4-verifiable-presentations-1_0.html)
 //! [JWT VC Presentation Profile]: (https://identity.foundation/jwt-vc-presentation-profile)
 
-pub mod client;
 pub mod dcql;
-mod endpoint;
+
 pub mod provider;
 pub mod types;
 pub mod vp_token;
 
+mod endpoint;
 mod error;
-mod handlers;
 mod state;
 
 use std::fmt::Display;
