--- src/oid4vp/client.rs
@@ -1,4 +1,0 @@
-//! # Client
-
-pub mod request;
-pub mod response;
