--- src/oid4vp/client/response.rs
@@ -1,1 +1,0 @@
-//! Response Builder
