--- src/oid4vp/handlers.rs
@@ -1,4 +1,0 @@
-mod create_request;
-mod metadata;
-mod request_uri;
-mod response;
