+++ src/oid4vp/provider.rs
@@ -3,12 +3,11 @@
 use std::future::Future;
 
 use anyhow::{Result, anyhow};
-use chrono::{DateTime, Utc};
 use credibil_identity::IdentityResolver;
 pub use credibil_identity::SignerExt;
-use serde::{Deserialize, Serialize};
 
-use crate::BlockStore;
+use crate::core::blockstore::BlockStore;
+pub use crate::core::state::StateStore;
 use crate::oid4vp::types::{Verifier, Wallet};
 use crate::token_status::StatusToken;
 
@@ -39,26 +38,8 @@ pub trait Metadata: Send + Sync {
     fn register(&self, verifier: &Verifier) -> impl Future<Output = Result<Verifier>> + Send;
 }
 
-/// `StateStore` is used to store and retrieve server state between requests.
-pub trait StateStore: Send + Sync {
-    /// Store state using the provided key. The expiry parameter indicates
-    /// when data can be expunged from the state store.
-    fn put(
-        &self, key: &str, state: impl Serialize + Send, expiry: DateTime<Utc>,
-    ) -> impl Future<Output = Result<()>> + Send;
-
-    /// Retrieve data using the provided key.
-    fn get<T>(&self, key: &str) -> impl Future<Output = Result<T>> + Send
-    where
-        T: for<'de> Deserialize<'de>;
-
-    /// Remove data using the key provided.
-    fn purge(&self, key: &str) -> impl Future<Output = Result<()>> + Send;
-}
-
 const WALLET: &str = "WALLET";
 const VERIFIER: &str = "VERIFIER";
-const STATE: &str = "STATE";
 
 impl<T: BlockStore> Metadata for T {
     async fn verifier(&self, verifier_id: &str) -> Result<Verifier> {
@@ -82,30 +63,5 @@ impl<T: BlockStore> Metadata for T {
         let block = serde_json::to_vec(&verifier)?;
         BlockStore::put(self, "owner", VERIFIER, &verifier.oauth.client_id, &block).await?;
         Ok(verifier)
-    }
-}
-
-impl<T: BlockStore> StateStore for T {
-    #[allow(unused)]
-    async fn put(
-        &self, key: &str, state: impl Serialize + Send, expiry: DateTime<Utc>,
-    ) -> Result<()> {
-        let state = serde_json::to_vec(&state)?;
-        BlockStore::delete(self, "owner", STATE, key).await?;
-        BlockStore::put(self, "owner", STATE, key, &state).await
-    }
-
-    async fn get<S>(&self, key: &str) -> Result<S>
-    where
-        S: for<'de> Deserialize<'de>,
-    {
-        let Some(block) = BlockStore::get(self, "owner", STATE, key).await? else {
-            return Err(anyhow!("could not find client"));
-        };
-        Ok(serde_json::from_slice(&block)?)
-    }
-
-    async fn purge(&self, key: &str) -> Result<()> {
-        BlockStore::delete(self, "owner", STATE, key).await
     }
 }
