+++ tests/vci_deferred.rs
@@ -5,14 +5,14 @@ use std::sync::LazyLock;
 
 use credibil_identity::{Key, SignerExt};
 use credibil_jose::{JwsBuilder, Jwt, decode_jws};
-use credibil_vc::core::did_jwk;
+use credibil_vc::blockstore::BlockStore;
 use credibil_vc::oid4vci::types::{
     CreateOfferRequest, Credential, CredentialHeaders, CredentialRequest, CredentialResponse,
     Dataset, DeferredCredentialRequest, DeferredHeaders, NonceRequest, ProofClaims, TokenGrantType,
     TokenRequest, W3cVcClaims,
 };
 use credibil_vc::oid4vci::{self, JwtType};
-use credibil_vc::{BlockStore, OneMany};
+use credibil_vc::{OneMany, did_jwk};
 use provider::issuer::{CAROL_ID, ISSUER_ID, Issuer, data};
 use provider::wallet::Wallet;
 use serde_json::json;
